requirements.txt
aiogram
aiohttp
python-dotenv

